feat: List VODs.

+++ server.ts
@@ -1,14 +1,33 @@
-const path = require('path')
-const express = require('express')
-const compression = require('compression')
-const morgan = require('morgan')
-const { createRequestHandler } = require('@remix-run/express')
+import { createRequestHandler } from '@remix-run/express'
+import compression from 'compression'
+import { config } from 'dotenv'
+import express, { Request, Response, NextFunction } from 'express'
+import morgan from 'morgan'
+import path from 'path'
+import serverTiming from 'server-timing'
+
+import { VodService } from './app/services/vods.server'
+
+config()
 
 const BUILD_DIR = path.join(process.cwd(), 'build')
+
+const vodService = new VodService()
+
+async function updateVodsLoop(): Promise<void> {
+  while (true) {
+    await vodService.updateVods()
+
+    await new Promise((resolve) => setTimeout(resolve, 1000 * 60 * 5))
+  }
+}
+
+updateVodsLoop()
 
 const app = express()
 
 app.use(compression())
+app.use(serverTiming())
 
 // http://expressjs.com/en/advanced/best-practice-security.html#at-a-minimum-disable-x-powered-by-header
 app.disable('x-powered-by')
@@ -25,22 +44,17 @@ app.use(express.static('public', { maxAge: '1h' }))
 
 app.use(morgan('tiny'))
 
-app.all(
-  '*',
-  process.env.NODE_ENV === 'development'
-    ? (req, res, next) => {
-        purgeRequireCache()
+app.all('*', (request: Request, response: Response, next: NextFunction) => {
+  if (process.env.NODE_ENV === 'development') {
+    purgeRequireCache()
+  }
 
-        return createRequestHandler({
-          build: require(BUILD_DIR),
-          mode: process.env.NODE_ENV,
-        })(req, res, next)
-      }
-    : createRequestHandler({
-        build: require(BUILD_DIR),
-        mode: process.env.NODE_ENV,
-      })
-)
+  return createRequestHandler({
+    build: require(BUILD_DIR),
+    mode: process.env.NODE_ENV
+  })(request, response, next)
+})
+
 const port = process.env.PORT || 3000
 
 app.listen(port, () => {
